+++ Sources/Player/Errors.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import Foundation
+
+extension Player {
+    enum PlaybackError: Error {
+        case unknown
+    }
+}
